requirements.txt
django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
drf-yasg==1.20.0
django-cors-headers==3.13.0
django-extensions==3.1.5
celery==5.2.7
pyjwt==2.4.0
django-filter==21.1
django-debug-toolbar==3.4.0
reportlab==3.6.10
xhtml2pdf==0.2.7
django-crispy-forms==1.13.0
django-redis==5.2.0
django-simple-history==2.12.0
django-cacheops==5.1
dateutils==0.6.12
python-dateutil==2.8.2
requests==2.27.1
sendgrid==6.9.4
boto3==1.20.43
firebase_admin==5.2.0
APScheduler==3.10.1